+++ rick_vfs/__init__.py
@@ -1,2 +1,2 @@
 from .version import get_version
-from .vfs import VfsObjectInfo, VfsVolume, VfsContainer, VfsError
+from .vfs import VfsObjectInfo, VfsVolume, VfsContainer, VfsError, VfsLockableVolume, VfsObjectLock